rusk: refactor config to remove grpc
